Moving EventDispatcher to utils folder
